requirements.txt
python-dotenv
git+https://github.com/hakutakaid/pyrogram.git
motor
pykeyboard
sqlalchemy==1.3.23
pyromod==1.5
